requirements.txt
numpy==1.21.4
paho-mqtt==1.6.1
Pillow==9.0.0
pynput==1.7.6
pyobjc-core==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Quartz==8.5
six==1.16.0
